requirements.txt
bcrypt==3.2.0
click == 8.1.2
pandas~=1.4.2
starlette~=0.20.0
uvicorn~=0.17.6
PyJWT~=2.4.0
SQLAlchemy~=1.4.37